Move `FeatureFlags` to `stripe-core`
